interacting with paradigm project
aded function getLogs to get the logs as cell array 200*3 where each ine is an event (beep). first column is the syllable said, second is the time the beep was played and third column is now wmpty and can be filled with time stamps vector.
additionly created a log file for our program.
all other functions support getLogFile().

**this commit should still be tested!!!